#8 refactor : annotation, exception 클래스를 common 모듈로 리팩토링 한다
