+++ api-server/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     implementation project(':domain')
     implementation project(':application')
     implementation project(':core')
+    implementation project(':common')
 
     // Spring Boot Starter
     implementation 'org.springframework.boot:spring-boot-starter-web'
