+++ api-server/src/main/java/kusitms/duduk/apiserver/healthcheck/presentation/HealthCheckControllerDocs.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import kusitms.duduk.core.exception.ErrorResponse;
+import kusitms.duduk.common.exception.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
