+++ api-server/src/main/java/kusitms/duduk/apiserver/security/presentation/OAuthControllerDocs.java
@@ -8,7 +8,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kusitms.duduk.core.security.dto.response.OAuthLoginResponse;
-import kusitms.duduk.core.exception.ErrorResponse;
+import kusitms.duduk.common.exception.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
