+++ application/build.gradle
@@ -10,6 +10,8 @@ jar.enabled = true
 dependencies {
     implementation(project(":domain"))
     implementation(project(":core"))
+    implementation project(':common')
+
 
     // Spring Data JPA
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
