+++ application/src/main/java/kusitms/duduk/application/attendant/persistence/AttendantPersistenceAdapter.java
@@ -2,7 +2,7 @@ package kusitms.duduk.application.attendant.persistence;
 
 import java.time.LocalDate;
 import kusitms.duduk.application.attendant.persistence.entity.AttendantJpaEntity;
-import kusitms.duduk.core.annotation.Adapter;
+import kusitms.duduk.common.annotation.Adapter;
 import kusitms.duduk.core.attendant.port.output.LoadAttendantPort;
 import kusitms.duduk.core.attendant.port.output.SaveAttendantPort;
 import lombok.RequiredArgsConstructor;
