+++ batch/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     implementation project(':application')
     implementation project(':domain')
     implementation project(':core')
+    implementation project(':common')
 
     // Spring Batch
     implementation 'org.springframework.boot:spring-boot-starter-batch'
