+++ common/build.gradle
@@ -1,19 +1,15 @@
 plugins {
     id 'org.springframework.boot' version '3.2.5'
     id 'io.spring.dependency-management' version '1.1.4'
+    id 'java-test-fixtures'
 }
 
 bootJar.enabled = false
 jar.enabled = true
 
 dependencies {
-    implementation(project(":domain"))
-    implementation project(':common')
-
     // Spring Boot Starter
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    // Swagger
-    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.2.0'
 }
 
 tasks.named('test') {
