+++ common/src/main/java/kusitms/duduk/common/annotation/DomainService.java
@@ -1,4 +1,4 @@
-package kusitms.duduk.core.annotation;
+package kusitms.duduk.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
