+++ common/src/main/java/kusitms/duduk/common/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package kusitms.duduk.core.exception;
+package kusitms.duduk.common.exception;
 
 public record ErrorResponse(
     int code,
