+++ common/src/main/java/kusitms/duduk/common/exception/custom/AlreadyExistsException.java
@@ -1,4 +1,4 @@
-package kusitms.duduk.core.exception.custom;
+package kusitms.duduk.common.exception.custom;
 
 public class AlreadyExistsException extends BaseException {
     public AlreadyExistsException(final String message) {
