+++ domain/build.gradle
@@ -8,7 +8,7 @@ bootJar.enabled = false
 jar.enabled = true
 
 dependencies {
-    implementation(project(":core"))
+    implementation project(':common')
 }
 
 tasks.named('test') {
