+++ domain/src/main/java/kusitms/duduk/domain/user/service/UpdateUserDomainService.java
@@ -1,13 +1,11 @@
 package kusitms.duduk.domain.user.service;
 
-import kusitms.duduk.core.annotation.DomainService;
-import kusitms.duduk.core.user.port.input.UpdateUserDomainUseCase;
+import kusitms.duduk.common.annotation.DomainService;
 import kusitms.duduk.domain.user.User;
 
 @DomainService
-public class UpdateUserDomainCommand implements UpdateUserDomainUseCase {
+public class UpdateUserDomainService {
 
-    @Override
     public void updateRefreshToken(User user, String refreshToken) {
         user.updateRefreshToken(refreshToken);
     }
